requirements.txt
annotated-types==0.6.0
anyio==3.7.1
attrs==23.1.0
blinker==1.7.0
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
coloredlogs==15.0.1
distro==1.8.0
exceptiongroup==1.1.3
Flask==3.0.0
flatbuffers==23.5.26
gTTS==2.4.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
humanfriendly==10.0
idna==3.4
imageio==2.32.0
imutils==0.5.4
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
lazy_loader==0.3
llvmlite==0.41.1
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.2.1
numba==0.58.1
numpy==1.26.2
onnxruntime==1.16.2
openai==1.3.2
opencv-python==4.8.1.78
opencv-python-headless==4.8.1.78
packaging==23.2
Pillow==10.0.1
platformdirs==4.0.0
pooch==1.8.0
progressbar==2.5
protobuf==4.25.1
pydantic==2.5.1
pydantic_core==2.14.3
PyMatting==1.1.11
referencing==0.31.0
rembg==2.0.52
requests==2.31.0
rpds-py==0.13.0
scikit-image==0.22.0
scipy==1.11.3
sniffio==1.3.0
sympy==1.12
tifffile==2023.9.26
tqdm==4.66.1
typing_extensions==4.8.0
urllib3==2.1.0
Werkzeug==3.0.1
